Updated Avatar badges and tweaked some things in profiles page
